+++ CHANGELOG.md
@@ -18,6 +18,7 @@ To see compatibility information, `xsvutils help compatibility`.
 
 以下のオプションが追加された。
 - grep コマンドの -v
+- uriparams コマンドの --sjis
 
 uriparamsの入力に `?...` や `http://.../...?...` の形式を許容するように変更した。
 
@@ -81,6 +82,7 @@ insconst のパラメータの順序が逆になってしまった。 (see `xsvu
 
 ### version 0.7 (-v7) (Unreleased)
 - add `--remove` option of `cols` subcommand
+- add `--sjis` option of `uriparams` subcommand
 
 ### version 0.6 (-v6) (2018/08/06)
 - add `-v` option of `grep` subcommand